refactor: restructure recommender into package
