+++ .gitignore
@@ -1,3 +1,5 @@
 /main.py
 /ml-25m/
 .idea/
+__pycache__/
+*.py[cod]
