+++ get_movie_ids.py
@@ -1,20 +1,11 @@
-import argparse
-import pandas as pd
-from paths import DEFAULT_MOVIES_PATH
+from pathlib import Path
+import sys
 
 
-def parse_args():
-    parser = argparse.ArgumentParser(description="Preview MovieLens movie IDs.")
-    parser.add_argument("--movies-path", default=DEFAULT_MOVIES_PATH, help="Path to movies.csv.")
-    parser.add_argument("--rows", type=int, default=5, help="Number of rows to display.")
-    return parser.parse_args()
+sys.path.insert(0, str(Path(__file__).resolve().parent / "src"))
 
-
-def main():
-    args = parse_args()
-    data = pd.read_csv(args.movies_path)
-    print(data.head(args.rows))
+from recommender.recommend import preview_movie_ids_main  # noqa: E402
 
 
 if __name__ == "__main__":
-    main()
+    preview_movie_ids_main()
