+++ src/recommender/__init__.py
@@ -1,0 +1,24 @@
+"""MovieLens recommender package."""
+
+__all__ = [
+    "MatrixFactorization",
+    "MovieLens",
+    "MovieRecommender",
+    "SimilarMovies",
+]
+
+
+def __getattr__(name):
+    if name == "MatrixFactorization":
+        from recommender.model import MatrixFactorization
+
+        return MatrixFactorization
+    if name == "MovieLens":
+        from recommender.data import MovieLens
+
+        return MovieLens
+    if name in {"MovieRecommender", "SimilarMovies"}:
+        from recommender.recommend import MovieRecommender, SimilarMovies
+
+        return {"MovieRecommender": MovieRecommender, "SimilarMovies": SimilarMovies}[name]
+    raise AttributeError(f"module 'recommender' has no attribute {name!r}")
